requirements.txt
qtmodern~=0.2.0
pyinstaller~=6.11.1
altgraph~=0.17.4
av~=14.0.1
packaging~=24.2
pillow~=11.1.0
PySide6~=6.8.1.1
xmlschema~=3.4.3
gpxpy~=1.6.2